add iteration questions
