+++ src/App.jsx
@@ -1,27 +1,14 @@
 import React, { useState } from 'react';
 import { Box, VStack, useColorModeValue } from '@chakra-ui/react';
-import QuestionCard from './components/QuestionCard';
 import DarkModeToggle from './components/DarkModeToggle';
 import TopicSelection from './components/TopicSelection';
-import quizData from './data/quadraticSequences.json';
+import QuizComponent from './components/QuizComponent';
 
 const App = () => {
-  const [currentQuestionIndex, setCurrentQuestionIndex] = useState(0);
   const [selectedTopic, setSelectedTopic] = useState(null);
-
-  const handleNextQuestion = () => {
-    if (currentQuestionIndex < quizData.questions.length - 1) {
-      setCurrentQuestionIndex(currentQuestionIndex + 1);
-    } else {
-      // Quiz finished logic here
-      console.log("Quiz finished!");
-      setSelectedTopic(null); // Reset to topic selection
-    }
-  };
 
   const handleSelectTopic = (topic) => {
     setSelectedTopic(topic);
-    setCurrentQuestionIndex(0); // Reset question index when new topic is selected
   };
 
   const bgColor = useColorModeValue('gray.50', 'gray.900');
@@ -32,13 +19,7 @@ const App = () => {
       <DarkModeToggle />
       <VStack spacing={8}>
         {selectedTopic ? (
-          <>
-            <h1>{selectedTopic} Quiz</h1>
-            <QuestionCard 
-              question={quizData.questions[currentQuestionIndex]}
-              onNext={handleNextQuestion}
-            />
-          </>
+          <QuizComponent topic={selectedTopic} />
         ) : (
           <TopicSelection onSelectTopic={handleSelectTopic} />
         )}
